# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk Sistem Data Induk Siswa SMK",
        "title": "API Siswa Induk",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@siswa.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/catatan-semester/{catatan_id}/pkl": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add internship record to semester notes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catatan Semester"
                ],
                "summary": "Add PKL to semester notes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Semester Notes ID",
                        "name": "catatan_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PKL data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreatePKLRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.PKLResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/mata-pelajaran": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of all active subjects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mata Pelajaran"
                ],
                "summary": "Get all subjects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/responses.MataPelajaranResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/siswa/{id}/kehadiran": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get attendance records for a student with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Siswa"
                ],
                "summary": "Get attendance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/responses.KehadiranResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/siswa/{id}/nilai-semester": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get semester grades for a student with filters and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nilai"
                ],
                "summary": "Get semester grades",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Class filter (X, XI, XII)",
                        "name": "kelas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Semester filter (1, 2)",
                        "name": "semester",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Academic year filter",
                        "name": "tahun_pelajaran",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/responses.NilaiSemesterResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/siswa/{siswa_id}/catatan-semester": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get semester notes for a student",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catatan Semester"
                ],
                "summary": "Get semester notes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "siswa_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/responses.CatatanSemesterResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create semester notes for a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catatan Semester"
                ],
                "summary": "Create semester notes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "siswa_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Semester notes data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateCatatanSemesterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.CatatanSemesterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/siswa/{siswa_id}/nilai-ijazah": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get certificate grades for a student",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nilai"
                ],
                "summary": "Get certificate grades",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "siswa_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/responses.NilaiIjazahResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a certificate grade for a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nilai"
                ],
                "summary": "Create certificate grade",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "siswa_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Certificate grade data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateNilaiIjazahRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.NilaiIjazahResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.CreatePKLRequest": {
            "type": "object",
            "required": [
                "nama_dudi"
            ],
            "properties": {
                "keterangan": {
                    "type": "string",
                    "example": "Praktek sebagai teknisi jaringan"
                },
                "lama_bulan": {
                    "type": "integer",
                    "example": 3
                },
                "lokasi": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "Bandung"
                },
                "nama_dudi": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "PT. Teknologi Indonesia"
                }
            }
        },
...
        "responses.KehadiranResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jumlah_alpa": {
                    "type": "integer"
                },
                "jumlah_hadir": {
                    "type": "integer"
                },
                "jumlah_hari_efektif": {
                    "type": "integer"
                },
                "jumlah_izin": {
                    "type": "integer"
                },
                "jumlah_sakit": {
                    "type": "integer"
                },
                "kelas": {
                    "type": "string"
                },
                "persentase_hadir": {
                    "type": "number"
                },
                "semester": {
                    "type": "integer"
                }
            }
        },
...
        "responses.MataPelajaranResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "kelompok": {
                    "type": "string"
                },
                "kode": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "sub_kelompok": {
                    "type": "string"
                }
            }
        },
...
        "responses.PKLResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "keterangan": {
                    "type": "string"
                },
                "lama_bulan": {
                    "type": "integer"
                },
                "lokasi": {
                    "type": "string"
                },
                "nama_dudi": {
                    "type": "string"
                }
            }
        },
...
        "utils.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/utils.Pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authenticate",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}